+++ hydra-node/test/Hydra/BehaviorSpec.hs
@@ -32,6 +32,7 @@ import Hydra.Chain (
   ChainEvent (..),
   ChainSlot (ChainSlot),
   ChainStateType,
+  HeadId (HeadId),
   HeadParameters (..),
   IsChainState,
   OnChainTx (..),
@@ -88,7 +89,7 @@ spec = parallel $ do
         withSimulatedChainAndNetwork $ \chain ->
           withHydraNode aliceSk [] chain $ \n1 -> do
             send n1 Init
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
             send n1 (Commit (utxoRef 1))
             waitUntil [n1] $ Committed alice (utxoRef 1)
 
@@ -97,7 +98,7 @@ spec = parallel $ do
         withSimulatedChainAndNetwork $ \chain ->
           withHydraNode aliceSk [] chain $ \n1 -> do
             send n1 Init
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
             send n1 (Commit (utxoRef 1))
             waitUntil [n1] $ Committed alice (utxoRef 1)
             waitUntil [n1] $ HeadIsOpen (utxoRef 1)
@@ -109,7 +110,7 @@ spec = parallel $ do
         withSimulatedChainAndNetwork $ \chain ->
           withHydraNode aliceSk [] chain $ \n1 -> do
             send n1 Init
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
             send n1 (Commit (utxoRef 1))
             waitUntil [n1] $ Committed alice (utxoRef 1)
             waitUntil [n1] $ HeadIsOpen (utxoRef 1)
@@ -121,7 +122,7 @@ spec = parallel $ do
         withSimulatedChainAndNetwork $ \chain ->
           withHydraNode aliceSk [] chain $ \n1 -> do
             send n1 Init
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
             send n1 (Commit (utxoRef 1))
             waitUntil [n1] $ Committed alice (utxoRef 1)
             waitUntil [n1] $ HeadIsOpen (utxoRef 1)
@@ -135,7 +136,7 @@ spec = parallel $ do
         withSimulatedChainAndNetwork $ \chain ->
           withHydraNode aliceSk [] chain $ \n1 -> do
             send n1 Init
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
             send n1 (Commit (utxoRef 1))
             waitUntil [n1] $ Committed alice (utxoRef 1)
             waitUntil [n1] $ HeadIsOpen (utxoRef 1)
@@ -152,7 +153,7 @@ spec = parallel $ do
           withHydraNode aliceSk [bob] chain $ \n1 ->
             withHydraNode bobSk [alice] chain $ \n2 -> do
               send n1 Init
-              waitUntil [n1, n2] $ ReadyToCommit (fromList [alice, bob])
+              waitUntil [n1, n2] $ HeadIsInitializing testHeadId (fromList [alice, bob])
 
               send n1 (Commit (utxoRef 1))
               waitUntil [n1] $ Committed alice (utxoRef 1)
@@ -169,13 +170,13 @@ spec = parallel $ do
           withHydraNode aliceSk [bob] chain $ \n1 ->
             withHydraNode bobSk [alice] chain $ \n2 -> do
               send n1 Init
-              waitUntil [n1, n2] $ ReadyToCommit (fromList [alice, bob])
+              waitUntil [n1, n2] $ HeadIsInitializing testHeadId (fromList [alice, bob])
               send n1 (Commit (utxoRefs [1, 2]))
               waitUntil [n1, n2] $ Committed alice (utxoRefs [1, 2])
               send n2 Abort
               waitUntil [n1, n2] $ HeadIsAborted (utxoRefs [1, 2])
               send n1 Init
-              waitUntil [n1, n2] $ ReadyToCommit (fromList [alice, bob])
+              waitUntil [n1, n2] $ HeadIsInitializing testHeadId (fromList [alice, bob])
 
     it "cannot abort head when commits have been collected" $
       shouldRunInSim $ do
@@ -183,7 +184,7 @@ spec = parallel $ do
           withHydraNode aliceSk [bob] chain $ \n1 ->
             withHydraNode bobSk [alice] chain $ \n2 -> do
               send n1 Init
-              waitUntil [n1, n2] $ ReadyToCommit (fromList [alice, bob])
+              waitUntil [n1, n2] $ HeadIsInitializing testHeadId (fromList [alice, bob])
               send n1 (Commit (utxoRef 1))
               send n2 (Commit (utxoRef 2))
 
@@ -198,7 +199,7 @@ spec = parallel $ do
           withHydraNode aliceSk [bob] chain $ \n1 ->
             withHydraNode bobSk [alice] chain $ \n2 -> do
               send n1 Init
-              waitUntil [n1, n2] $ ReadyToCommit (fromList [alice, bob])
+              waitUntil [n1, n2] $ HeadIsInitializing testHeadId (fromList [alice, bob])
 
               send n1 (Commit (utxoRef 1))
               waitUntil [n1] $ Committed alice (utxoRef 1)
@@ -219,7 +220,7 @@ spec = parallel $ do
             withHydraNode aliceSk [bob] chain $ \n1 ->
               withHydraNode bobSk [alice] chain $ \n2 -> do
                 send n1 Init
-                waitUntil [n1, n2] $ ReadyToCommit (fromList [alice, bob])
+                waitUntil [n1, n2] $ HeadIsInitializing testHeadId (fromList [alice, bob])
                 send n1 (Commit (utxoRef 1))
 
                 waitUntil [n2] $ Committed alice (utxoRef 1)
@@ -408,7 +409,7 @@ spec = parallel $ do
             withSimulatedChainAndNetwork $ \chain ->
               withHydraNode aliceSk [] chain $ \n1 -> do
                 send n1 Init
-                waitUntil [n1] $ ReadyToCommit (fromList [alice])
+                waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
                 send n1 (Commit (utxoRef 1))
 
           logs = selectTraceEventsDynamic @_ @(HydraNodeLog SimpleTx) result
@@ -423,13 +424,13 @@ spec = parallel $ do
             withSimulatedChainAndNetwork $ \chain ->
               withHydraNode aliceSk [] chain $ \n1 -> do
                 send n1 Init
-                waitUntil [n1] $ ReadyToCommit (fromList [alice])
+                waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
                 send n1 (Commit (utxoRef 1))
 
           logs = selectTraceEventsDynamic @_ @(HydraNodeLog SimpleTx) result
 
-      logs `shouldContain` [BeginEffect alice (ClientEffect $ ReadyToCommit $ fromList [alice])]
-      logs `shouldContain` [EndEffect alice (ClientEffect $ ReadyToCommit $ fromList [alice])]
+      logs `shouldContain` [BeginEffect alice (ClientEffect $ HeadIsInitializing testHeadId $ fromList [alice])]
+      logs `shouldContain` [EndEffect alice (ClientEffect $ HeadIsInitializing testHeadId $ fromList [alice])]
 
     roundtripAndGoldenSpecs (Proxy @(HydraNodeLog SimpleTx))
 
@@ -439,18 +440,18 @@ spec = parallel $ do
         withSimulatedChainAndNetwork $ \chain ->
           withHydraNode aliceSk [] chain $ \n1 -> do
             send n1 Init
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
             -- We expect the Init to be rolled back and forward again
             rollbackAndForward chain 1
             waitUntil [n1] RolledBack
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
 
     it "does work for rollbacks past open" $
       shouldRunInSim $ do
         withSimulatedChainAndNetwork $ \chain ->
           withHydraNode aliceSk [] chain $ \n1 -> do
             send n1 Init
-            waitUntil [n1] $ ReadyToCommit (fromList [alice])
+            waitUntil [n1] $ HeadIsInitializing testHeadId (fromList [alice])
             send n1 (Commit (utxoRef 1))
             waitUntil [n1] $ Committed alice (utxoRef 1)
             waitUntil [n1] $ HeadIsOpen (utxoRefs [1])
@@ -643,7 +644,7 @@ createMockNetwork node nodes =
 toOnChainTx :: UTCTime -> PostChainTx tx -> OnChainTx tx
 toOnChainTx now = \case
   InitTx HeadParameters{contestationPeriod, parties} ->
-    OnInitTx{contestationPeriod, parties}
+    OnInitTx{contestationPeriod, parties, headId = testHeadId}
   (CommitTx pa ut) ->
     OnCommitTx pa ut
   AbortTx{} ->
@@ -665,6 +666,9 @@ toOnChainTx now = \case
 -- NOTE(SN): Deliberately long to emphasize that we run these tests in IOSim.
 testContestationPeriod :: ContestationPeriod
 testContestationPeriod = UnsafeContestationPeriod 3600
+
+testHeadId :: HeadId
+testHeadId = HeadId "1234"
 
 nothingHappensFor ::
   (MonadTimer m, MonadThrow m, IsChainState tx) =>
@@ -751,7 +755,7 @@ openHead ::
   IOSim s ()
 openHead n1 n2 = do
   send n1 Init
-  waitUntil [n1, n2] $ ReadyToCommit (fromList [alice, bob])
+  waitUntil [n1, n2] $ HeadIsInitializing testHeadId (fromList [alice, bob])
   send n1 (Commit (utxoRef 1))
   waitUntil [n1, n2] $ Committed alice (utxoRef 1)
   send n2 (Commit (utxoRef 2))